+++ lib/osm.dart
@@ -8,6 +8,7 @@ export 'src/filter.dart';
 export 'src/pbf/exception.dart';
 export 'src/pbf/file.dart';
 export 'src/pbf/header.dart';
+export 'src/pbf/writer.dart';
 export 'src/xml/change.dart';
 export 'src/xml/exception.dart';
 export 'src/subset.dart';